split login and register a bit
